backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Vector Database
chromadb==0.4.18

# Numpy compatibility (chromadb 0.4.18 needs numpy < 2.0)
numpy<2.0

# HTTP Client (async)
httpx==0.25.2

# Authentication & Security
PyJWT==2.10.1
python-jose==3.3.0

# Rate Limiting
slowapi==0.1.9

# Web Scraping
beautifulsoup4==4.12.2
feedparser==6.0.10

# Date/Time Utilities
python-dateutil==2.8.2

# AI/Summarization (using Ollama - local LLM)
# Note: httpx is already included above for Ollama API calls

# Logging
python-json-logger==2.0.7

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
ruff==0.1.8
mypy==1.7.1
